create file system create src/database src/queries, changed view directory to public, add reset.css file relates #3
